Split accuracy unit tests for norm operators
