+++ tests/test_skip_layer_norm.py
@@ -1,0 +1,45 @@
+import pytest
+import torch
+
+import flag_gems
+
+from . import accuracy_utils as utils
+from . import conftest as cfg
+
+if cfg.QUICK_MODE:
+    FLOAT_DTYPES = [torch.float32]
+else:
+    FLOAT_DTYPES = utils.FLOAT_DTYPES
+
+
+@pytest.mark.skip_layer_norm
+@pytest.mark.parametrize("shape", utils.REDUCTION_SHAPES)
+@pytest.mark.parametrize("dtype", FLOAT_DTYPES)
+def test_skip_layernorm(shape, dtype):
+    N = shape[1]
+    layer_shape = [
+        N,
+    ]
+    inp = torch.randn(shape[:2], dtype=dtype, device=flag_gems.device)
+    residual = torch.randn(shape[:2], dtype=dtype, device=flag_gems.device)
+    weight = torch.randn(layer_shape, dtype=dtype, device=flag_gems.device)
+    bias = torch.randn(layer_shape, dtype=dtype, device=flag_gems.device)
+    eps = 1e-5
+
+    ref_inp = utils.to_reference(inp, True)
+    ref_residual = utils.to_reference(residual, True)
+    ref_weight = utils.to_reference(weight, True)
+    ref_bias = utils.to_reference(bias, True)
+
+    ref_out = torch.layer_norm(
+        ref_inp + ref_residual,
+        list(layer_shape),
+        weight=ref_weight,
+        bias=ref_bias,
+        eps=eps,
+    )
+    res_out = flag_gems.skip_layer_norm(
+        inp, residual, list(layer_shape), weight=weight, bias=bias, eps=eps
+    )
+
+    utils.gems_assert_close(res_out, ref_out, dtype)
